Move requests to restAPI package to fix coverage
